requirements.txt
boto3==1.18.6
python-dotenv==0.19.0
requests==2.26.0